fix: accent fixes, greeting logic, specialization multi-select, profile extraction from tests
- Fix Spanish accents across multiple components (años, aparecerán, únicamente)
- Fix greeting logic: show "Buenas noches" between 0-6h instead of "Buenos días"
- RegisterPsychologist: specialization field changed to multi-select dropdown with chips
- UserSettingsTab: re-fetch profile from server after save instead of optimistic update
- UserConsentTab: optimistic update after signing + month format long
- CalendarWeek: month short → long on mobile
- Backend: extract age, birthDate, chiefComplaint from initial/matching test answers

+++ frontend/src/components/CalendarWeek.tsx
@@ -657,7 +657,7 @@ export default function CalendarWeek({ mode, slots, myAppointments = [], onCreat
             <div className="text-center">
               {isMobile ? (
                 <span className="font-heading text-base font-bold text-gantly-text tracking-tight">
-                  {visibleDays[0]?.toLocaleDateString('es-ES', { day: 'numeric', month: 'short' })} &ndash; {visibleDays[visibleDays.length - 1]?.toLocaleDateString('es-ES', { day: 'numeric', month: 'short' })}
+                  {visibleDays[0]?.toLocaleDateString('es-ES', { day: 'numeric', month: 'long' })} &ndash; {visibleDays[visibleDays.length - 1]?.toLocaleDateString('es-ES', { day: 'numeric', month: 'long' })}
                 </span>
               ) : (
                 <span className="font-heading text-xl font-bold text-gantly-text tracking-tight">
